% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igascores.R
\name{igascores}
\alias{igascores}
\title{Generate various scores for IgA binding in IgA-Seq experiments}
\usage{
igascores(
  posabunds = NULL,
  negabunds = NULL,
  possizes = NULL,
  negsizes = NULL,
  pseudo = NULL,
  totabunds = NULL,
  method = "probratio",
  scaleratio = TRUE
)
}
\arguments{
\item{posabunds}{A dataframe of taxa abundances in the positive/high IgA gate samples. Samples as columns and taxa as rows, column and row names must match across abundance tables.}

\item{negabunds}{A dataframe of taxa abundances in the negative/low IgA gate samples. Samples as columns and taxa as rows, column and row names must match across abundance tables.}

\item{possizes}{A named vector containing the fraction of events in the IgA postive gate for each sample, with sample names matching abundance dataframes.}

\item{negsizes}{A named vector containing the fraction of events in the IgA negative gate for each sample, with sample names matching abundance dataframes.}

\item{pseudo}{The pseudo count to be used in scores. Default is 1e-5. Recommend setting to minimum observed abundance.}

\item{totabunds}{A dataframe of taxa abundances in the whole/initial samples. Samples as columns and taxa as rows, column and row names must match across abundance tables.}

\item{method}{Method to use to score IgA binding. One of: "probratio","prob","kau","palm". Default is "probratio".}

\item{scaleratio}{Should probratio scores be scaled to the pseudocount. Default is TRUE.}
}
\description{
This function enables calculation of a variety of different indicies for scoring IgA binding to taxa in IgA-Seq experiments.
It is designed to be called on dataframes of abundance values, allowing easy calculation of scores across multiple taxa and samples.
The igaprobabilityratio(), igaprobability(), kauindex() and palmindex() functions can be used to calculate scores for one taxa and one sample.

Scoring method can be chosen by specifying the method parameter as one of: "probratio", "prob", "kau", "palm" (Defaults to "probratio").
Each method requires different inputs as detailed below:

\itemize{
\item probratio - equivalent to igaprobabilityratio() - requires two seperate dataframes with iga positive abundances and iga negative abundances, two vectors with the sizes of the iga postive and negative gates per sample, and a pseudo count
\item prob - equivalent to igaprobability() - requires a dataframe with iga pos or neg fraction abundances, a vector of iga pos or neg gate size per sample, and a dataframe of taxa abudnances in the whole/inital samples
\item kau - equivalent to kauindex() - requires two seperate dataframes with iga positive abundances and iga negative abundances, and a pseudo count
\item palm - equivalent to palmindex() - requires two seperate dataframes with iga positive abundances and iga negative abundances, and a pseudo count
}
}
\keyword{Jackson,}
\keyword{Kau,}
\keyword{Palm,}
\keyword{experiment,}
\keyword{iga,}
\keyword{iga-seq}
\keyword{index,}
\keyword{probability,}
\keyword{ratio,}
\keyword{score,}
