% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oligoMeta_data.R
\docType{data}
\name{oligoMeta}
\alias{oligoMeta}
\title{Metadata for the OligoMM12-Colitis experiment used as an example in the IgAScores package}
\format{
An object of class \code{"tibble"}.
}
\usage{
data(oligoMeta)
}
\description{
Metadata associated with the oligoSpecies data set.
}
\details{
Metadata for an experiment where mice with a defined gut microbiota (OligoMM12) were either given *Helicobacter hepaticus* and IL10R antibody or the antibody alone (the first developing colitis).
These data accompany the species level counts in oligoSpecies and also include a negative extraction control. A subset of the combined condition were not properly colonised by *H.hepaticus* and excluded from later analyses.
Further details can be found in Jackson et al. (2020, \doi{10.1101/2020.08.19.257501}).
}
\examples{
data(oligoMeta)
}
\references{
To come...
}
\keyword{ail10r}
\keyword{colitis}
\keyword{dataset}
\keyword{helicobacter}
\keyword{metadata}
\keyword{oligoMM12}
