% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relabund.R
\name{relabund}
\alias{relabund}
\title{Relative Abundance from Counts}
\usage{
relabund(counttable, percentage = FALSE)
}
\arguments{
\item{counttable}{Dataframe with rows as observations and columns as samples.}

\item{percentage}{Should values be returned as a percentage? i.e multiplied by 100. Default is FALSE (as required for most IgA scoring approaches).}
}
\description{
This function converts values in a dataframe to a fraction/percentage of the sum of their column.
}
\keyword{abundance,}
\keyword{microbiome}
\keyword{normalisation,}
