% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxnamesplit.R
\name{taxnamesplit}
\alias{taxnamesplit}
\title{Extract name at a given taxonomic level from a full name in the "p__;c__;o__;f__;g__;s__" format}
\usage{
taxnamesplit(names, level = "genus")
}
\arguments{
\item{names}{Name string/ vector of name strings}

\item{level}{taxonomic level to split at must be in range phylum to species (default is genus).}
}
\value{
A string (or vector of strings if input is a vector) containing the second part of the input string split at the given taxonomic level.
}
\description{
This function splits a full taxonomic lineage as a given level and returns the latter half.
}
\examples{
taxnamesplit("p__Bacteroidetes;c__Bacteroidia","class")
}
\keyword{microbiome}
\keyword{name}
\keyword{split}
\keyword{taxonomy}
